core: Add truncation indicator for bar with vertical filling
This adds two new attributes "column_size" and "column_size_max" to
bars, which allow to configure widths of columns when filling is set to
"columns_horizontal" or "columns_vertical". The new attributes function
similar to "size" and "size_max".

In case a bar item is truncated due to exceeding the column width, an
indicator (weechat.look.bar_more_line) is shown. The color of the
indicator is specified by weechat.color.bar_more.